+++ include/nc_runtime/nc_types.h
@@ -38,6 +38,7 @@
 #endif
 
 // single precision
+#define NC_TWO_PI 6.2831853f // 2 * pi
 #define NC_PI 3.14159265f    // pi
 #define NC_PI_2 1.57079633f  // pi/2
 #define NC_PI_4 0.785398163f // pi/4
@@ -45,11 +46,12 @@
 #define NC_2_PI 0.636619772f // 2/pi
 
 // double precision
-#define NC_PI_D 3.14159265358979323846    // pi
-#define NC_PI_2_D 1.57079632679489661923  // pi/2
-#define NC_PI_4_D 0.785398163397448309616 // pi/4
-#define NC_1_PI_D 0.318309886183790671538 // 1/pi
-#define NC_2_PI_D 0.636619772367581343076 // 2/pi
+#define NC_TWO_PI_D 6.28318530717958647692 // 2 * pi
+#define NC_PI_D 3.14159265358979323846     // pi
+#define NC_PI_2_D 1.57079632679489661923   // pi/2
+#define NC_PI_4_D 0.785398163397448309616  // pi/4
+#define NC_1_PI_D 0.318309886183790671538  // 1/pi
+#define NC_2_PI_D 0.636619772367581343076  // 2/pi
 
 // clang-format off
 static forceinline float nc_radianToDegree(float radian) { return radian * (180.0f / NC_PI); }